requirements.txt
python-docx==0.8.11
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7 